+++ packages/driver/src/cy/commands/navigation.ts
@@ -35,9 +35,8 @@ const reset = (test: any = {}) => {
   // before each test run!
   previouslyVisitedLocation = undefined
 
-  // make sure we reset that we haven't visited about blank again
-  // strict test isolation resets the navigation history for us.
-  hasVisitedAboutBlank = Cypress.config('testIsolation') === 'on'
+  // TODO: clean this up and remove any unnecessary logic
+  hasVisitedAboutBlank = true
 
   currentlyVisitingAboutBlank = false