+++ system-tests/projects/session-and-origin-e2e-specs/cypress/e2e/session/new_session_and_fails_validation_retries.cy.js
@@ -1,0 +1,11 @@
+/**
+ * Used in cy-in-cy tests in @packages/app.
+ */
+it('t1', { retries: 1 }, () => {
+  const setupFn = cy.stub()
+  const validateFn = cy.stub().throws(new Error('validation error'))
+
+  cy.session('blank_session', setupFn, {
+    validate: validateFn,
+  })
+})